requirements.txt
boto3
botocore
moto[ec2,s3]
Flask
python-dotenv